Add adjustments to VLM to make them work without dtype warnings

+++ inference_experimental/inference_exp/models/paligemma/paligemma_hf.py
@@ -64,7 +64,7 @@ class PaliGemmaHF:
             base_model_path = os.path.join(model_name_or_path, "base")
             model = PaliGemmaForConditionalGeneration.from_pretrained(
                 base_model_path,
-                torch_dtype=torch_dtype,
+                dtype=torch_dtype,
                 trust_remote_code=trust_remote_code,
                 local_files_only=local_files_only,
                 quantization_config=quantization_config,
@@ -83,7 +83,7 @@ class PaliGemmaHF:
         else:
             model = PaliGemmaForConditionalGeneration.from_pretrained(
                 model_name_or_path,
-                torch_dtype=torch_dtype,
+                dtype=torch_dtype,
                 device_map=device,
                 trust_remote_code=trust_remote_code,
                 local_files_only=local_files_only,
